requirements.txt
requests
bs4
selenium
pillow
pymysql
pandas
openpyxl
